api/requirements.txt
boto3
pandas
mysql-connector-python
fastapi
requests
uvicorn
uuid